+++ src/main/java/ch/heigvd/amt/projet/shop_els/controller/ArticleAddController.java
@@ -1,4 +1,6 @@
 package ch.heigvd.amt.projet.shop_els.controller;
+
+import ch.heigvd.amt.projet.shop_els.model.Category;
 import ch.heigvd.amt.projet.shop_els.util.HibUtil;
 import org.hibernate.Session;
 
@@ -9,14 +11,28 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
+import java.io.PrintWriter;
 import java.util.List;
 
-@WebServlet("/admin/articles")
-public class Article extends HttpServlet{
+@WebServlet("/admin/articleAdd")
+public class ArticleAddController extends HttpServlet {
     private Session session;
 
     @Override
     protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+        response.setContentType("text/html");
+        session = HibUtil.getSessionFactory().getCurrentSession();
+        session.beginTransaction();
+        Query query = session.getNamedQuery("selectAllCategory");
+        List<Category> results = query.getResultList();
+        session.close();
+
+        request.setAttribute("categories", results);
+        request.getRequestDispatcher("/WEB-INF/view/admin/articleAdd.jsp").forward(request, response);
+    }
+
+    @Override
+    protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
 
         response.setContentType("text/html");
         session = HibUtil.getSessionFactory().getCurrentSession();
